+++ Graphics/Renderable.cs
@@ -1,0 +1,6 @@
+namespace ProjectTerra.Framework.Graphics;
+
+public interface IRenderable
+{
+    public int GetHandle();
+}
